deletes and edits and sweet alerts working

+++ clientsrc/src/Pages/BugDetails.vue
@@ -13,37 +13,57 @@
       <p class="mb-1">Created By:</p>
       <h5>{{profile.name}}</h5>
     </div>
-    <div class="col-6 card text-center pt-3 shadow-lg px-0">
+    <div class="col-6 card p-2 border-primary text-center pt-3 shadow-lg px-0">
+      <i
+        type="button"
+        v-if="bug.creatorEmail == $auth.user.email && bug.closed==false"
+        data-toggle="modal"
+        data-target="#edit-modal"
+        class="fa fa-pencil text-right px-4"
+      ></i>
+      <editModal :editBugData="bug" />
       <h2>{{bug.description}}</h2>
       <button
         type="button"
-        class="btn btn-outline-danger btn-block rounded"
+        class="btn btn-outline-danger btn-block rounded mt-3"
         @click="deleteBug(bug.id)"
       >Resolve Bug Issue</button>
     </div>
-    <div class="card shadow-lg mt-4 col-10">
+    <div class="card border-primary shadow-lg my-4 p-3 col-10">
       <button
+        v-if="bug.closed==true"
+        disabled
         type="button"
         data-toggle="modal"
         data-target="#note-modal"
-        class="btn btn-primary mt-3 m-2 shadow-lg"
+        class="btn btn-outline-primary mt-3 m-2 shadow-lg"
+      >Add A Note</button>
+      <button
+        v-else
+        type="button"
+        data-toggle="modal"
+        data-target="#note-modal"
+        class="btn btn-outline-primary mt-3 m-2 shadow-lg"
       >Add A Note</button>
       <noteModal />
-      <div class="p-2" v-for="note in notes" :key="note.id">
+      <div class="p-2 overflow" v-for="note in notes" :key="note.id">
+        <hr />
         <button
           type="button"
           class="btn btn-outline-danger py-0 px-1"
-          @click="deleteNote(notes[id])"
+          @click="deleteNote(note.id)"
         >X</button>
         : {{note.content}}
       </div>
     </div>
+    <footer class="col-12 bg-primary text-center text-white py-3">Created By: Kayden Cooper</footer>
   </div>
 </template>
 
 
 <script>
 import NoteModal from "../components/NoteModal";
+import EditModal from "../components/EditModal";
 export default {
   name: "bugDetails",
   data() {
@@ -66,19 +86,51 @@ export default {
   },
   methods: {
     deleteBug(id) {
-      this.$store.dispatch("deleteBug", id);
+      swal({
+        title: "Are you sure?",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+      }).then((willDelete) => {
+        if (willDelete) {
+          this.$store.dispatch("deleteBug", id);
+          swal("Successfully Deleted!", {
+            icon: "success",
+          });
+        } else {
+          swal("Phew Close One!");
+        }
+      });
     },
     deleteNote(id) {
-      debugger;
-      this.$store.dispatch("deleteNote", id);
+      swal({
+        title: "Are you sure?",
+
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+      }).then((willDelete) => {
+        if (willDelete) {
+          this.$store.dispatch("deleteNote", id);
+          swal("Successfully Deleted!", {
+            icon: "success",
+          });
+        } else {
+          swal("Phew Close One!");
+        }
+      });
     },
   },
   components: {
     NoteModal,
+    EditModal,
   },
 };
 </script>
 
 
 <style scoped>
+.overflow {
+  overflow: auto;
+}
 </style>
